requirements.txt
streamlit==1.28.0
fastapi==0.104.0
uvicorn==0.24.0
pypdf2==3.0.1
python-docx==0.8.11
chromadb==0.4.15
sentence-transformers==2.2.2
requests==2.31.0
python-multipart==0.0.6
